Extract sharing super types from unions
